pcb coil code for minimum turns.py
this code has a logical error the equation for inductance in nano henry is given. to find the minimum outer diameter we need to give the equation in first derivative test the clearance nd the widht of trace is given the only variable present is turns. now with least no of turns we get least outer dia. but the error we get is the turns get to -ve which solves the outer diameter with complex number.